requirements.txt
appdirs
attrs
beautifulsoup4
blinker==1.5
Brotli
bs4
certifi
cffi
charset-normalizer
click
colorama
cryptography
cssselect
dnspython
fake-useragent
Flask
flask-cors
gunicorn
h11
h2
hpack
hyperframe
idna
importlib_metadata
itsdangerous
Jinja2
kaitaistruct
lxml
lxml_html_clean
MarkupSafe
outcome
packaging
parse
psutil
pyasn1
pycparser
pydivert
pyee
PyJWT
pymongo
pyOpenSSL
pyparsing
pyppeteer
pyquery
PySocks
python-dotenv
pytz
requests
requests-html
selenium
selenium-wire
setuptools
sniffio
sortedcontainers
soupsieve
termcolor
tqdm
trio
trio-websocket
typing_extensions
undetected-chromedriver
urllib3
w3lib
webdriver-manager
websocket-client
websockets
Werkzeug
wsproto
yt-dlp
zipp
zstandard

google-api-python-client==2.179.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-core
cachetools
httplib2
uritemplate